refactor: establish homeboy-observation-contract crate keystone
Creates homeboy-observation-contract and seeds it with the pure artifact-record
type cluster: ArtifactRecord + ArtifactViewerLink (+ the default_artifact_type
serde fn). These are behavior-free serde types describing a recorded run
artifact; they're carried by consumers like the extension bench report layer,
which shouldn't have to depend on the observation store just for the record
vocabulary.

- New crate deps only serde/serde_json (pure leaf below core).
- homeboy-core re-exports the types (crate::observation::{ArtifactRecord,
  ArtifactViewerLink}) so existing call sites are unchanged. ArtifactRecord's
  only nested type is ArtifactViewerLink (moves with it); everything else is
  primitives.
- extension bench/artifact.rs (which carries Vec<ArtifactViewerLink>) is FULLY
  severed from crate::observation. bench/artifact_persistence.rs takes
  ArtifactRecord from the contract too, keeping only its genuine ActiveObservation
  / finding-records behavior deps.

Extension's observation edge drops from 13 files to 12. Same keystone-first
discipline as refactor-contract (#8774): build the seam below core, drain type
edges over time toward a future homeboy-extension extraction.

Refs #8425
